More safe mechanism against invalid creations of datasets; renamed "Groups" to "Datasets"; removed interface "Verifiable"

+++ src/main/java/io/github/siemieniuk/votingsystems/ballot/entry/CumulativeBallotEntry.java
@@ -7,9 +7,9 @@ import java.util.Objects;
 
 @Setter
 @Getter
-public class CumulativeBallotEntry extends EntryWithNumber<CandidateEntry, Float> {
+public class CumulativeBallotEntry extends EntryWithNumber<CandidateEntry, Double> {
 
-    public CumulativeBallotEntry(CandidateEntry preference, float fraction) {
+    public CumulativeBallotEntry(CandidateEntry preference, double fraction) {
         super(preference, fraction);
     }
